+++ CHANGELOG.md
@@ -6,9 +6,10 @@ This file contains the list of changes made to the Joulescope driver.
 
 ## 1.1.3
 
-2023 Jan 20
+2023 Jan 24
 
 * Fixed open to correctly handle error on lower-level device open.
+* Improved statistics output to include time/sample_freq and time/range. 
 
 
 ## 1.1.2